% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{report_html}
\alias{report_html}
\title{Render NIBSC HTML report}
\usage{
report_html(...)
}
\arguments{
\item{...}{Arguments to be passed to the function `[bookdown::html_document2]`}
}
\value{
A modified `html_document2` with specific parameters for NIBSC reports
}
\description{
This function is called in the output of the YAML header of the .Rmd file used to generate the report
to specify the standard output desired as html
}
\examples{
\dontrun{
   output: nibscdoc::report_html
}
}
