% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{report_word}
\alias{report_word}
\title{Render NIBSC Word report}
\usage{
report_word(...)
}
\arguments{
\item{...}{Arguments to be passed to the function `[bookdown::word_document2]`}
}
\value{
A modified `word_document2` with specific parameters for NIBSC reports
}
\description{
This function is called in the output of the YAML header of the .Rmd file used to generate the report
to specify the standard output desired as Word document
}
\examples{
#' \dontrun{
   output: nibscdoc::report_word
}
}
